# cmd/api/openapi.yaml
openapi: 3.0.3
info:
  title: Scutum API
  version: 1.0.0
  description: |
    Scutum is a self-hosted infrastructure orchestration platform.
    It manages containers (Docker), Kubernetes deployments, WireGuard networking,
    S3-compatible storage, Git sync, plugin extensions, and emergency KMS recovery.

    ## Authentication
    Most endpoints require authentication. Two methods are accepted:
    - **JWT Bearer token** — obtained via `POST /auth/login`
    - **API Key** — created via `POST /auth/keys`, sent in the `X-API-Key` header
servers:
  - url: http://localhost:8080
    description: Local development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from `POST /auth/login`
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key obtained from `POST /auth/keys`
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: invalid credentials
        code:
          type: string
          example: UNAUTHORIZED
        trace_id:
          type: string
          example: abc123
        time:
          type: string
          format: date-time
...
    UpdateRoleRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        perms:
          type: array
          items:
            type: string

    # ── Nodes ─────────────────────────────────────────────────────────────────
    NodeRecord:
      type: object
      properties:
        id:
          type: string
          example: n_core01
        name:
          type: string
          example: core-01
        type:
          type: string
          enum: [hub, edge, hub+edge]
          example: hub
        address:
          type: string
          example: 10.0.0.1:51820
        public_key:
          type: string
          example: abc123==
    CreateNodeRequest:
      type: object
      required: [name, type, address, public_key]
      properties:
        name:
          type: string
          example: core-01
        type:
          type: string
          enum: [hub, edge, hub+edge]
          example: hub
        address:
          type: string
          example: 10.0.0.1:51820
        public_key:
          type: string
          example: abc123==

    # ── Sync ─────────────────────────────────────────────────────────────────
...
  responses:
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Authenticated but insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BadRequest:
      description: Invalid request body or parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /roles/{id}:
    put:
      tags: [Roles]
      summary: Update a role
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRoleRequest'
      responses:
        '204':
          description: Updated
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Roles]
      summary: Delete a role
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Setup ─────────────────────────────────────────────────────────────────
...
  /nodes:
    get:
      tags: [Nodes]
      summary: List mesh nodes
      responses:
        '200':
          description: Node list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NodeRecord'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Nodes]
      summary: Register a node
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateNodeRequest'
      responses:
        '201':
          description: Node registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodeRecord'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...